+++ helpers.ts
@@ -1,10 +1,11 @@
-import { PROGRAM_ID, TreeConfig } from "@metaplex-foundation/mpl-bubblegum";
+import { PROGRAM_ID, TreeConfig, TokenProgramVersion, TokenStandard } from "@metaplex-foundation/mpl-bubblegum";
 import {
   Connection,
   PublicKey,
   Signer,
   Transaction,
   TransactionInstruction,
+  AccountMeta
 } from "@solana/web3.js";
 import { BN, Provider } from "@project-serum/anchor";
 import { PROGRAM_ID as TOKEN_METADATA_PROGRAM_ID } from "@metaplex-foundation/mpl-token-metadata";
@@ -39,7 +40,7 @@ export async function execute(
   provider: Provider,
   instructions: TransactionInstruction[],
   signers: Signer[],
-  skipPreflight = false,
+  skipPreflight = true,
   verbose = false
 ): Promise<string> {
   let tx = new Transaction();
@@ -50,8 +51,10 @@ export async function execute(
   let txid: string | null = null;
   try {
     txid = await provider.sendAndConfirm!(tx, signers, {
+      commitment: "confirmed",
       skipPreflight,
     });
+
   } catch (e: any) {
     console.log("Tx error!", e.logs);
     throw e;
@@ -108,3 +111,39 @@ export async function getMasterEdition(mint: PublicKey) {
     )
   )[0];
 }
+
+export const mapProof = (assetProof: { proof: string[] }): AccountMeta[] => {
+  if (!assetProof.proof || assetProof.proof.length === 0) {
+    throw new Error("Proof is empty");
+  }
+  return assetProof.proof.map((node) => ({
+    pubkey: new PublicKey(node),
+    isSigner: false,
+    isWritable: false,
+  }));
+};
+
+export const makeCompressedNFT = (
+  name: string,
+  symbol: string,
+  uri: string
+) => {
+  return {
+    name: name,
+    symbol: symbol,
+    uri: uri,
+    creators: [],
+    editionNonce: 253,
+    tokenProgramVersion: TokenProgramVersion.Original,
+    tokenStandard: TokenStandard.NonFungible,
+    uses: null,
+    collection: null,
+    primarySaleHappened: false,
+    sellerFeeBasisPoints: 0,
+    isMutable: false,
+  };
+};
+
+export const sleep = async (ms: any) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+};
