+++ lib/create-tree.ts
@@ -1,0 +1,58 @@
+import { execute } from "../helpers";
+import { Keypair, PublicKey, SystemProgram } from "@solana/web3.js";
+import { WrappedConnection } from "../wrappedConnection";
+import {
+  createCreateTreeInstruction,
+  PROGRAM_ID as BUBBLEGUM_PROGRAM_ID,
+} from "@metaplex-foundation/mpl-bubblegum";
+import {
+  SPL_ACCOUNT_COMPRESSION_PROGRAM_ID,
+  getConcurrentMerkleTreeAccountSize,
+  SPL_NOOP_PROGRAM_ID,
+} from "@solana/spl-account-compression";
+
+export const createTree = async (
+  connectionWrapper: WrappedConnection,
+  payerKeypair: Keypair,
+  maxDepth: number = 14,
+  maxBufferSize: number = 64
+) => {
+  const payer = payerKeypair.publicKey;
+  const merkleTreeKeypair = Keypair.generate();
+  const merkleTree = merkleTreeKeypair.publicKey;
+  const space = getConcurrentMerkleTreeAccountSize(maxDepth, maxBufferSize, 5);
+  const allocTreeIx = SystemProgram.createAccount({
+    fromPubkey: payer,
+    newAccountPubkey: merkleTree,
+    lamports: await connectionWrapper.getMinimumBalanceForRentExemption(space),
+    space: space,
+    programId: SPL_ACCOUNT_COMPRESSION_PROGRAM_ID,
+  });
+  const [treeAuthority, _bump] = await PublicKey.findProgramAddress(
+    [merkleTree.toBuffer()],
+    BUBBLEGUM_PROGRAM_ID
+  );
+  const createTreeIx = createCreateTreeInstruction(
+    {
+      merkleTree,
+      treeAuthority,
+      treeCreator: payer,
+      payer,
+      logWrapper: SPL_NOOP_PROGRAM_ID,
+      compressionProgram: SPL_ACCOUNT_COMPRESSION_PROGRAM_ID,
+    },
+    {
+      maxBufferSize,
+      maxDepth,
+      public: false,
+    },
+    BUBBLEGUM_PROGRAM_ID
+  );
+  await execute(
+    connectionWrapper.provider,
+    [allocTreeIx, createTreeIx],
+    [merkleTreeKeypair, payerKeypair]
+  );
+
+  return merkleTree;
+};
