+++ package.json
@@ -24,7 +24,7 @@
     "@metaplex-foundation/solita": "^0.18.0",
     "@project-serum/anchor": "0.26.0",
     "@solana/spl-account-compression": "^0.1.4",
-    "@solana/spl-token": "^0.1",
+    "@solana/spl-token": "^0.3.7",
     "@solana/web3.js": "1.73.0",
     "@sorend-solana/utils": "^0.0.5",
     "borsh": "^0.7.0",
